requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
appdirs==1.4.4
apturl==0.5.2
asgiref==3.5.0
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.1
bcrypt==3.2.0
blinker==1.6.2
Brlapi==0.8.3
cachelib==0.10.2
cachetools==5.3.1
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.2.0
chrome-gnome-shell==0.0.0
click==8.1.6
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dataclasses-json==0.5.14
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==3.2.12
docker-pycreds==0.4.0
duplicity==0.8.21
faiss-gpu==1.7.2
fasteners==0.14.1
filelock==3.6.0
Flask==2.3.2
Flask-Mail==0.9.1
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
frozenlist==1.4.0
future==0.18.2
gitdb==4.0.10
GitPython==3.1.36
greenlet==2.0.2
gTTS==2.3.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jwt==1.3.1
keyring==23.5.0
langchain==0.0.260
langsmith==0.0.20
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
mysqlclient==1.4.6
netifaces==0.11.0
numexpr==2.8.5
numpy==1.25.2
oauthlib==2.1.0
olefile==0.46
openai==0.28.1
openapi-schema-pydantic==1.2.4
opencv-python==4.8.0.76
packaging==23.1
pandas==2.1.0
paramiko==2.9.3
pathtools==0.1.2
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
protobuf==4.21.12
psutil==5.9.6
ptyprocess==0.7.0
pyarrow==13.0.0
pycairo==1.20.1
pycups==2.0.1
pydantic==1.10.12
pydeck==0.8.0
pygame==2.5.1
pyglet==2.0.9
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.1.0
pymacaroons==0.13.0
Pympler==1.0.1
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
PyNaCl==1.5.0
pyparsing==2.4.7
PyPDF2==3.0.1
pyRFC3339==1.1
pytesseract==0.3.10
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==5.4.1
referencing==0.30.2
regex==2023.8.8
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.1.0
rich==13.5.3
rpds-py==0.10.3
scikit-learn==1.3.0
scipy==1.11.2
SecretStorage==3.3.1
sentry-sdk==1.32.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.19
sqlparse==0.4.2
streamlit==1.26.0
systemd-python==234
tenacity==8.2.2
threadpoolctl==3.2.0
tiktoken==0.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.7
usb-creator==0.3.7
validators==0.22.0
virtualenv==20.13.0+ds
wadllib==1.3.6
Wand==0.6.11
wandb==0.15.12
watchdog==3.0.0
Werkzeug==2.3.6
WTForms==3.0.1
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0